add a new page for onchain-graph
