+++ src/pages/OnchainGraph/types/nft.ts
